chore: Bump Rust toolchain to 1.90

+++ crates/q_cli/src/cli/doctor/mod.rs
@@ -272,15 +272,15 @@ where
     let args = args.into_iter().collect::<Vec<_>>();
 
     Some(DoctorFix::Sync(Box::new(move || {
-        if let (Some(exe), Some(remaining)) = (args.first(), args.get(1..)) {
-            if Command::new(exe).args(remaining).status()?.success() {
-                if let Some(duration) = sleep_duration.into() {
-                    let spinner = Spinner::new(Spinners::Dots, "Waiting for command to finish...".into());
-                    std::thread::sleep(duration);
-                    stop_spinner(Some(spinner)).ok();
-                }
-                return Ok(());
+        if let (Some(exe), Some(remaining)) = (args.first(), args.get(1..))
+            && Command::new(exe).args(remaining).status()?.success()
+        {
+            if let Some(duration) = sleep_duration.into() {
+                let spinner = Spinner::new(Spinners::Dots, "Waiting for command to finish...".into());
+                std::thread::sleep(duration);
+                stop_spinner(Some(spinner)).ok();
             }
+            return Ok(());
         }
         eyre::bail!(
             "Failed to run {:?}",
@@ -335,10 +335,10 @@ fn print_status_result(name: impl Display, status: &Result<(), DoctorError>, ver
                     println!("  {infoline}");
                 }
             }
-            if let Some(error) = error {
-                if verbose {
-                    println!("  {error:?}");
-                }
+            if let Some(error) = error
+                && verbose
+            {
+                println!("  {error:?}");
             }
         },
     }
@@ -466,18 +466,18 @@ impl DoctorCheck for DesktopSocketCheck {
         let fig_socket_path = directories::desktop_socket_path().context("No socket path")?;
         let parent = fig_socket_path.parent().map(PathBuf::from);
 
-        if let Some(parent) = parent {
-            if !parent.exists() {
-                return Err(DoctorError::Error {
-                    reason: format!("{PRODUCT_NAME} socket parent directory does not exist").into(),
-                    info: vec![format!("Path: {}", fig_socket_path.display()).into()],
-                    fix: Some(DoctorFix::Sync(Box::new(|| {
-                        std::fs::create_dir_all(parent)?;
-                        Ok(())
-                    }))),
-                    error: None,
-                });
-            }
+        if let Some(parent) = parent
+            && !parent.exists()
+        {
+            return Err(DoctorError::Error {
+                reason: format!("{PRODUCT_NAME} socket parent directory does not exist").into(),
+                info: vec![format!("Path: {}", fig_socket_path.display()).into()],
+                fix: Some(DoctorFix::Sync(Box::new(|| {
+                    std::fs::create_dir_all(parent)?;
+                    Ok(())
+                }))),
+                error: None,
+            });
         }
 
         check_file_exists(&fig_socket_path).map_err(|_err| {
@@ -831,28 +831,24 @@ impl DoctorCheck for PtySocketCheck {
                         ..
                     },
                 )) => {
-                    if let Some(style) = zsh_autosuggestion_style {
-                        if let Some(fg) = style.fg {
-                            if let Some(fig_proto::figterm::term_color::Color::Indexed(i)) = fg.color {
-                                if i == 15 {
-                                    return Err(doctor_warning!(
-                                        "ZSH_AUTOSUGGEST_HIGHLIGHT_STYLE is set to the same style as your text, this must be different to detect what you've typed"
-                                    ));
-                                }
-                            }
-                        }
+                    if let Some(style) = zsh_autosuggestion_style
+                        && let Some(fg) = style.fg
+                        && let Some(fig_proto::figterm::term_color::Color::Indexed(i)) = fg.color
+                        && i == 15
+                    {
+                        return Err(doctor_warning!(
+                            "ZSH_AUTOSUGGEST_HIGHLIGHT_STYLE is set to the same style as your text, this must be different to detect what you've typed"
+                        ));
                     }
 
-                    if let Some(style) = fish_suggestion_style {
-                        if let Some(fg) = style.fg {
-                            if let Some(fig_proto::figterm::term_color::Color::Indexed(i)) = fg.color {
-                                if i == 15 {
-                                    return Err(doctor_warning!(
-                                        "The Fish suggestion color is set to the same style your text, this must be different in order to detect what you've typed"
-                                    ));
-                                }
-                            }
-                        }
+                    if let Some(style) = fish_suggestion_style
+                        && let Some(fg) = style.fg
+                        && let Some(fig_proto::figterm::term_color::Color::Indexed(i)) = fg.color
+                        && i == 15
+                    {
+                        return Err(doctor_warning!(
+                            "The Fish suggestion color is set to the same style your text, this must be different in order to detect what you've typed"
+                        ));
                     }
                 },
                 _ => {
@@ -898,10 +894,10 @@ impl DoctorCheck<Option<Shell>> for DotfileCheck {
     }
 
     async fn get_type(&self, current_shell: &Option<Shell>, _platform: Platform) -> DoctorCheckType {
-        if let Some(shell) = current_shell {
-            if *shell == self.integration.get_shell() {
-                return DoctorCheckType::NormalCheck;
-            }
+        if let Some(shell) = current_shell
+            && *shell == self.integration.get_shell()
+        {
+            return DoctorCheckType::NormalCheck;
         }
 
         if is_executable_in_path(self.integration.get_shell().to_string()) {
@@ -1380,12 +1376,12 @@ impl DoctorCheck<Option<Terminal>> for ItermIntegrationCheck {
     }
 
     async fn check(&self, _: &Option<Terminal>) -> Result<(), DoctorError> {
-        if let Some(version) = app_version("com.googlecode.iterm2") {
-            if version < Version::new(3, 4, 0) {
-                return Err(doctor_error!(
-                    "iTerm version is incompatible with {PRODUCT_NAME}. Please update iTerm to latest version"
-                ));
-            }
+        if let Some(version) = app_version("com.googlecode.iterm2")
+            && version < Version::new(3, 4, 0)
+        {
+            return Err(doctor_error!(
+                "iTerm version is incompatible with {PRODUCT_NAME}. Please update iTerm to latest version"
+            ));
         }
         Ok(())
     }
@@ -1567,10 +1563,10 @@ impl DoctorCheck<Option<Terminal>> for VSCodeIntegrationCheck {
                 let glob_set = glob([extensions.join("withfig.fig-").to_string_lossy()]).unwrap();
 
                 let extensions = extensions.as_path();
-                if let Ok(fig_extensions) = glob_dir(&glob_set, extensions) {
-                    if fig_extensions.is_empty() {
-                        missing = false;
-                    }
+                if let Ok(fig_extensions) = glob_dir(&glob_set, extensions)
+                    && fig_extensions.is_empty()
+                {
+                    missing = false;
                 }
             }
 
@@ -1882,10 +1878,11 @@ where
 
         let mut result = check.check(&context).await;
 
-        if !config.strict && check_type == DoctorCheckType::SoftCheck {
-            if let Err(DoctorError::Error { reason, .. }) = result {
-                result = Err(DoctorError::Warning(reason));
-            }
+        if !config.strict
+            && check_type == DoctorCheckType::SoftCheck
+            && let Err(DoctorError::Error { reason, .. }) = result
+        {
+            result = Err(DoctorError::Warning(reason));
         }
 
         if config.all || result.is_err() {
@@ -2012,10 +2009,10 @@ pub async fn doctor_cli(all: bool, strict: bool) -> Result<ExitCode> {
     }
 
     // Remove update lock on doctor runs to fix bad state if update crashed.
-    if let Ok(update_lock) = fig_util::directories::update_lock_path(&Context::new()) {
-        if update_lock.exists() {
-            std::fs::remove_file(update_lock).ok();
-        }
+    if let Ok(update_lock) = fig_util::directories::update_lock_path(&Context::new())
+        && update_lock.exists()
+    {
+        std::fs::remove_file(update_lock).ok();
     }
 
     launch_fig_desktop(LaunchArgs {
